+++ Model/Entities/Entity.cs
@@ -1,5 +1,4 @@
-﻿using Platform.Model.CRUD;
-using System;
+﻿using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.ComponentModel;
@@ -7,7 +6,7 @@ using System.Linq;
 
 namespace Platform.Model
 {
-    public abstract class Entity<TKey> : IEntity<TKey>, IRecord, INotifyPropertyChanged
+    public abstract class Entity<TKey> : IEntity<TKey>, INotifyPropertyChanged
 #if !PORTABLE
 , ICloneable, INotifyPropertyChanging
 #endif
@@ -64,38 +63,7 @@ namespace Platform.Model
         #endregion
 
         #region Operators
-
-        //public static bool operator ==(Entity<TKey> var1, Entity<TKey> var2)
-        //{
-
-        //    if (object.ReferenceEquals(var1, null))
-        //    {
-        //        return object.ReferenceEquals(var2, null);
-        //    }
-
-        //    if (object.ReferenceEquals(var2, null))
-        //    {
-        //        return object.ReferenceEquals(var1, null);
-        //    }
-
-        //    return var1.Id.Equals(var2.Id);
-
-        //}
-        //public static bool operator !=(Entity<TKey> var1, Entity<TKey> var2)
-        //{
-        //    if (object.ReferenceEquals(var1, null))
-        //    {
-        //        return object.ReferenceEquals(var2, null);
-        //    }
-
-        //    if (object.ReferenceEquals(var2, null))
-        //    {
-        //        return object.ReferenceEquals(var1, null);
-        //    }
-
-        //    return !var1.Id.Equals(var2.Id);
-        //}
-
+        
         public override bool Equals(Object obj)
         {
             // Check for null values and compare run-time types.
@@ -106,58 +74,9 @@ namespace Platform.Model
         }
 
         #endregion
-
-        #region Extended LSE
-
-        protected internal bool IsLoaded;
-
-        public event EventHandler Loaded;
-        public virtual void Load()
-        {
-            if (Loaded != null) { Loaded(this, new EventArgs()); }
-        }
-
-        public event EventHandler Saved;
-        public virtual void Save()
-        {
-            if (Saved != null) { Saved(this, new EventArgs()); }
-        }
-
-        public event EventHandler Erased;
-        public virtual void Erase()
-        {
-            if (Erased != null) { Erased(this, new EventArgs()); }
-        }
-
-        #endregion
-
-        #region Basic CRUD
-
-        public virtual object Create() { return null; }
-        public virtual object Read() { return null; }
-        public virtual object Update() { return null; }
-        public virtual object Delete() { return null; }
-
-        #endregion
-
+        
         public virtual void Refresh(object e) { }
-
-        #region EventHandler
-
-        public event EventHandler<RecordEventArgs> Changed;
-        public void OnChanged(RecordEventArgs e)
-        {
-            if (Changed != null) { Changed(this, e); }
-        }
-
-        public event EventHandler<RecordEventArgs> Error;
-        public void OnError(RecordEventArgs e)
-        {
-            if (Error != null) { Error(this, e); }
-        }
-
-        #endregion
-
+        
         public virtual IEnumerable AsEnumerable() { return null; }
         public override int GetHashCode()
         {
@@ -169,11 +88,11 @@ namespace Platform.Model
         public object Clone()
         {
             Type _ty = GetType();
-            Entity _object = (Entity)Activator.CreateInstance(_ty);
+            Entity<TKey> _object = (Entity<TKey>)Activator.CreateInstance(_ty);
             CopyTo(_object);
             return _object;
         }
-        public void CopyTo(Entity target)
+        public void CopyTo(Entity<TKey> target)
         {
             if (object.ReferenceEquals(GetType(), target.GetType()))
             {
@@ -202,15 +121,5 @@ namespace Platform.Model
 #endif
 
     }
-
-    public abstract class Entity : Entity<long>
-    {
-
-        public override object Create() { return Id == 0; }
-        public override object Read() { return Id == 0; }
-        public override object Update() { return Id != 0; }
-        public override object Delete() { return Id != 0; }
-
-    }
-
+    
 }