+++ Support.Data/Attributes/IgnoreAttribute.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace Platform.Support.Data.Attributes
-{
-    [AttributeUsage(AttributeTargets.Property)]
-    public class IgnoreAttribute : Attribute
-    {
-    }
-}